@Luke-zhang-04

- feat(bash): add quickstart ascii art 08983df
- feat(bash): add colours to log 3f02a5c
- feat(bash): check for `.gitignore` before adding it 0a46ec2 
- add new logo d679cf0
- refactor (bash): refactor(bash): move functions to seperate files as modules a8d634a
- refactor(bash): add colours.sh for colour valiables 4e1f014
- feat(bash): add eslint configuration ca7e641
- feat(bash/eslint): add more logs db379d9 
- feat(bash): add codeclimate feature bffeefe 
- feat(bash): add stylelint option 0160f0f
- update tests with artifact uploading eca54eb 
- feat(bash): add seperate make command for makefile generation b8622aa 
- feat(bash): add independent eslint installation if no react app found 5dd58ca 
- feat(bash): add more sophisticated parameter checking 3d9beb2

+++ bash/main.sh
@@ -1,86 +1,60 @@
 #!/usr/bin/env bash
+# Quickstart
+# Copyright (c) 2020 Luke Zhang | https://luke-zhang-04.github.io/ | MIT Licence
 
-args="$*" # Passed in arguments
+DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" >/dev/null 2>&1 && pwd )" # Get location of this script
+scripts=("bootstrap" "eslint" "react" "codeclimate" "stylelint" "make")
+
+# Import other scripts from bash directory
+for script in "${scripts[@]}"; do
+    . "${DIR}/${script}.sh"
+done
+
+. "${DIR}/colours.sh"
+
+argString="$*" # Passed in arguments
+IFS=" " read -ra args <<< "$argString"
+quickstartAscii="$Cyan
+  ____        _      _        _             _   
+ / __ \      (_)    | |      | |           | |  
+| |  | |_   _ _  ___| | _____| |_ __ _ _ __| |_ 
+| |  | | | | | |/ __| |/ / __| __/ _\` | '__| __|
+| |__| | |_| | | (__|   <\__ \ || (_| | |  | |_ 
+ \___\__\__,_|_|\___|_|\_\___/\__\__,_|_|   \__|
+"
+startupText="${Cyan}Check us out on GitHub! https://github.com/Luke-zhang-04/quickstart\nWe promise that none of our dependencies, remote repositories/Gists, or shell scripts contain malicious code, but why trust us? Go see for yourself!\nhttps://gist.github.com/Luke-zhang-04/140bea238fa6dec12929c220645540e1\nhttps://gist.github.com/Luke-zhang-04/d4c19d39f0a462fc79e0b3361752cf95\n"
 
 #######################################
-# Checks parameters
+# Checks parameters and returns their existence
 # Globals:
 #   none
 # Arguments:
-#   global to set to
 #   full arg name
 #   short arg name (optional)
 #######################################
 checkParams() { # Check passed in arguments and match them
-    if [[ "$args" == *"$2"* ]]; then
-        eval "$1"=true
-    elif [[ "$args" == *"$3"* ]]&&[[ "$3" ]]; then
-        eval "$1"=true
-    else 
-        eval "$1"=false
-    fi
+    for arg in "${args[@]}"; do
+        if [[ "$arg" == "$1" ]]; then
+            printf true
+            return 0
+        elif [[ "$arg" == "$2" ]]; then
+            printf true
+            return 0
+        fi
+    done
+    printf false
+    return 1
 }
 
 # Check parameters for all the variables
-checkParams react_app react
-checkParams typescript typescript ts
-checkParams eslint eslint
-checkParams bootstrap bootstrap bs
-checkParams noVer noVer nv
-
-cleanup() {
-    if ("$react"); then
-        cd .. # cd back 1 level
-    fi
-    rm -rf ./quickstart # Get rid of quickstart
-}
-
-#######################################
-# Create react app
-# Globals:
-#   bootstrap
-# Arguments:
-#   None
-#######################################
-getBootstrap() {
-    if "$bootstrap"; then # If Bootstrap was requested
-        printf "\tFound bootstrap\n\tInstalling Bootstrap...\n"
-        yarn add bootstrap # Add bootstrap with Yarn
-        printf "\tInstalled Boostrap!\n\tMaking SCSS directory...\n"
-        mkdir scss # Make a SCSS directory
-
-        printf "Adding default imports to scss/bootstrap.scss...\n"
-        # Add default imports into ./scss/bootstrap.scss
-        printf "/*!\n * Bootstrap v4.5.0 (https://getbootstrap.com/)\n * Copyright 2011-2020 The Bootstrap Authors\n * Copyright 2011-2020 Twitter, Inc.\n * Licensed under MIT (https://github.com/twbs/bootstrap/blob/master/LICENSE)\n */\n\n// Required\n@import \"./../node_modules/bootstrap/scss/functions\";\n@import \"./../node_modules/bootstrap/scss/variables\";\n@import \"./../node_modules/bootstrap/scss/mixins\";\n\n// Optional\n@import \"./../node_modules/bootstrap/scss/root\";\n@import \"./../node_modules/bootstrap/scss/reboot\";\n@import \"./../node_modules/bootstrap/scss/type\";\n@import \"./../node_modules/bootstrap/scss/images\";\n@import \"./../node_modules/bootstrap/scss/code\";\n@import \"./../node_modules/bootstrap/scss/grid\";\n@import \"./../node_modules/bootstrap/scss/tables\";\n@import \"./../node_modules/bootstrap/scss/forms\";\n@import \"./../node_modules/bootstrap/scss/buttons\";\n@import \"./../node_modules/bootstrap/scss/transitions\";\n@import \"./../node_modules/bootstrap/scss/dropdown\";\n@import \"./../node_modules/bootstrap/scss/button-group\";\n@import \"./../node_modules/bootstrap/scss/input-group\";\n@import \"./../node_modules/bootstrap/scss/custom-forms\";\n@import \"./../node_modules/bootstrap/scss/nav\";\n@import \"./../node_modules/bootstrap/scss/navbar\";\n@import \"./../node_modules/bootstrap/scss/card\";\n@import \"./../node_modules/bootstrap/scss/breadcrumb\";\n@import \"./../node_modules/bootstrap/scss/pagination\";\n@import \"./../node_modules/bootstrap/scss/badge\";\n@import \"./../node_modules/bootstrap/scss/jumbotron\";\n@import \"./../node_modules/bootstrap/scss/alert\";\n@import \"./../node_modules/bootstrap/scss/progress\";\n@import \"./../node_modules/bootstrap/scss/media\";\n@import \"./../node_modules/bootstrap/scss/list-group\";\n@import \"./../node_modules/bootstrap/scss/close\";\n@import \"./../node_modules/bootstrap/scss/toasts\";\n@import \"./../node_modules/bootstrap/scss/modal\";\n@import \"./../node_modules/bootstrap/scss/tooltip\";\n@import \"./../node_modules/bootstrap/scss/popover\";\n@import \"./../node_modules/bootstrap/scss/carousel\";\n@import \"./../node_modules/bootstrap/scss/spinners\";\n@import \"./../node_modules/bootstrap/scss/utilities\";\n@import \"./../node_modules/bootstrap/scss/print\";\n" >> scss/bootstrap.scss
-        printf "\tInstalled Bootstrap!\n"
-    else
-        printf "\tBootstrap not found\n"
-    fi
-}
-
-#######################################
-# Create react app
-# Globals:
-#   react_app
-#   typescript
-# Arguments:
-#   None
-#######################################
-reactApp() {
-    if "$react_app"&&"$typescript"; then # React app w/ typescript
-        printf "\tFound React with Typescript\n\tInstalling React with Typescript...\n"
-        npx create-react-app client --template typescript # Create React app named client, with typescript template
-        printf "\tInstalled React with Typescript!\n"
-        cd client # cd into client
-    elif "$react_app"; then # React app wo/ typescript
-        printf "\tFound React without Typescript\n\tInstalling React...\n"
-        npx create-react-app client # Create React app names client
-        printf "\tInstalled React!\n"
-        cd client # cd into client
-    else
-        printf "\tReact not found\n"
-    fi
-}
+bootstrap="$(checkParams bootstrap bs)"
+codeclimate="$(checkParams codeclimate cc)"
+eslint="$(checkParams eslint esl)"
+make="$(checkParams makefile make)"
+noVer="$(checkParams noVer nv)"
+typescript="$(checkParams typescript ts)"
+react_app="$(checkParams react)"
+stylelint="$(checkParams stylelint slint)"
 
 #######################################
 # Main function
@@ -93,42 +67,63 @@ reactApp() {
 #   None
 #######################################
 quickStart() {
-    # Clone the quickstart Gist. Check this Gist for malicious code first, in case you don't trust me (don't trust anyone).
-    printf "Cloning quickstart gist from https://gist.github.com/d4c19d39f0a462fc79e0b3361752cf95.git"
-    git clone https://gist.github.com/d4c19d39f0a462fc79e0b3361752cf95.git
-    mv d4c19d39f0a462fc79e0b3361752cf95 quickstart # Change name to quickstart
+    # Clone the quickstart Gist. Check this Gist's code first, in case you don't trust me (don't trust anyone).
+    printf "${IBlue}Cloning quickstart gist from https://gist.github.com/d4c19d39f0a462fc79e0b3361752cf95.git${Cyan}\n"
+    git clone https://gist.github.com/d4c19d39f0a462fc79e0b3361752cf95.git quickstart
 
-    printf "Getting .gitignore file...\n"
-    mv ./quickstart/.gitignore ./.gitignore # Get .gitignore from quickstart
+    printf "${IBlue}Checking for .gitignore file...${Cyan}\n"
+    if test -f ".gitignore"; then
+        printf "\t${IYellow}.gitignore exists${Cyan}\n"
+    else
+        printf "\t${IGreen}.gitignore does not exist${Cyan}\n"
+        printf "\t${IBlue}Getting .gitignore file...${Cyan}\n"
+        mv ./quickstart/.gitignore ./.gitignore # Get .gitignore from quickstart
+        printf "\t${IGreen}Got .gitignore file${Cyan}\n"
+    fi
 
-    printf "Checking for React...\n"
-    reactApp
+    printf "${IBlue}Checking for React...${Cyan}\n"
+    reactApp "$react_app" "$typescript"
 
-    printf "Checking for Bootstrap...\n"
-    getBootstrap
+    printf "${IBlue}Checking for Bootstrap...${Cyan}\n"
+    getBootstrap "$bootstrap"
 
-    printf "Cleaning up...\n"
-    cleanup
+    printf "${IBlue}Checking for Eslint...${Cyan}\n"
+    getEslint "$eslint" "$typescript" "$react_app"
+
+    printf "${IBlue}Checking for CodeClimate...${Cyan}\n"
+    getCodeClimate "$codeclimate"
+
+    printf "${IBlue}Checking for Stylelint...${Cyan}\n"
+    getStylelint "$stylelint"
+
+    printf "${IBlue}Checking for Make...${Cyan}\n"
+    makeMakefile "$make" "$bootstrap" "$eslint" "$stylelint"
+
+    printf "${IGreen}Cleaning up...${Cyan}\n"
+    rm -rf ./quickstart # Get rid of quickstart
 }
 
 # Ask for proceeding
-printf "Preparing to quickstart with: $args"
+printf "$quickstartAscii"
+printf "$startupText"
+printf "${Cyan}Preparing to quickstart with:${Purple} $argString${RESET}\n"
 
-if ! [ "$noVer" ]; then
-    read -n 1 -p "Proceed? [Y/n] " proceed
-else
+if $noVer; then
     proceed="Y"
+else
+    printf "${Yellow}Proceed? [${BIGreen}Y${Yellow}/${IRed}n${Yellow}]${RESET} "
+    read proceed
 fi
 
 if [ "$proceed" = "Y" ]; then
-    printf "\nProceeding with quickstart\n"
-    quickStart
-    printf "Quickstart succesfully quickstarted your project!"
+    printf "\n${Cyan}Proceeding with quickstart${Cyan}\n"
+    quickStart # Invoke main quickstart function
+    printf "${BIGreen}Quickstart succesfully quickstarted your project!"
     exit 0
 elif [ "$proceed" = "n" ]; then
-    printf "\nAborting quickstart . . .\n"
+    printf "\n${BIRed}Aborting quickstart . . .\n"
     exit 0
 else
-    printf "\nUnknown input $proceed, aborting\n"
+    printf "\n${BIRed}Unknown input $proceed, aborting\n"
     exit 1
 fi
